Fix precision problem

+++ src/core/lipp_base.h
@@ -12,20 +12,20 @@ class LinearModel
 {
 public:
     double a = 0; // slope
-    double b = 0; // intercept
+    long double b = 0; // intercept
 
     LinearModel() = default;
-    LinearModel(double a, double b) : a(a), b(b) {}
+    LinearModel(double a, long double b) : a(a), b(b) {}
     explicit LinearModel(const LinearModel &other) : a(other.a), b(other.b) {}
 
     inline int predict(T key) const
     {
-        return std::floor(a * static_cast<double>(key) + b);
+        return std::floor(a * static_cast<long double>(key) + b);
     }
 
     inline double predict_double(T key) const
     {
-        return a * static_cast<double>(key) + b;
+        return a * static_cast<long double>(key) + b;
     }
 };
